refactor: store 폴더를 하나로 통합합니다.
