+++ src/frontend/src/layouts/main-layout/index.tsx
@@ -11,7 +11,7 @@ import ServerIcon from '@/components/server-icon'
 import { statusKo } from '@/constants/status'
 import { useGetServer } from '@/hooks/queries/server/useGetServer'
 import useGetSelfUser from '@/hooks/queries/user/useGetSelfUser'
-import useChattingStomp from '@/hooks/store/use-chatting-stomp'
+import useChattingStomp from '@/stores/use-chatting-stomp'
 import { cn } from '@/libs/cn'
 import { useMediaSettingsStore } from '@/stores/use-media-setting.store'
 import { useSignalingStomp } from '@/stores/use-signaling-stomp-store'
