+++ src/frontend/src/pages/channel/components/voice.tsx
@@ -3,9 +3,9 @@ import { useEffect, useState } from 'react'
 import ChatArea from '@/components/chat-area'
 import CustomButton from '@/components/custom-button'
 import useGetSelfUser from '@/hooks/queries/user/useGetSelfUser'
-import { useSignalingWithSFU } from '@/hooks/store/use-signaling-with-sfu'
-import useUserStatus from '@/hooks/store/use-user-status'
+import { useSignalingWithSFU } from '@/hooks/use-signaling-with-sfu'
 import { cn } from '@/libs/cn'
+import useUserStatus from '@/stores/use-user-status'
 import { ChatUser } from '@/types/user'
 
 import VideoCard from './video-card'
